+++ betty/demo/serve.py
@@ -11,8 +11,8 @@ from betty.demo.generate import generate_with_cleanup
 from betty.demo.project import create_project
 from betty.job import Context
 from betty.locale.localizable.gettext import _
-from betty.plugins.server import builtin
 from betty.server import Server, ServerNotStarted
+from betty.servers import project_builtin
 
 if TYPE_CHECKING:
     from betty.app import App
@@ -48,7 +48,7 @@ class DemoServer(Server):
                 _("Generating site...")
             ) as progress:
                 await generate_with_cleanup(project, context=Context(progress=progress))
-            self._server = await builtin.Builtin.new(project)
+            self._server = await project_builtin.ProjectBuiltinServer.new(project)
             await self._exit_stack.enter_async_context(self._server)
         except BaseException:
             # __aexit__() is not called when __aenter__() raises an exception, so ensure we clean up our resources.
