+++ betty/tests/plugins/extension/http_api_doc/test_swagger_ui.py
@@ -4,10 +4,10 @@ import pytest
 from playwright.async_api import Page, expect
 
 from betty.plugins.extension.http_api_doc import HttpApiDoc
-from betty.plugins.server import builtin
 from betty.project import Project
 from betty.project.generate import generate
 from betty.server import Server
+from betty.servers import project_builtin
 from betty.tests.conftest import (
     check_skip_playwright,
     check_skip_webpack_entry_point_provider,
@@ -19,7 +19,9 @@ class TestSwaggerUi:
     async def served_project(self) -> AsyncIterator[tuple[Project, Server]]:
         async with Project.new_isolated(extensions=[HttpApiDoc]) as project:
             await generate(project)
-            async with await builtin.Builtin.new(project) as server:
+            async with await project_builtin.ProjectBuiltinServer.new(
+                project
+            ) as server:
                 yield project, server
 
     @pytest.mark.asyncio(loop_scope="session")
