+++ betty/tests/plugins/extension/raspberry_mint/test_search_ui.py
@@ -6,10 +6,10 @@ from playwright.async_api import Page, expect
 from betty.plugins.entity.person import Person
 from betty.plugins.entity.person_name import PersonName
 from betty.plugins.extension.raspberry_mint import RaspberryMint
-from betty.plugins.server import builtin
 from betty.project import Project
 from betty.project.generate import generate
 from betty.server import Server
+from betty.servers import project_builtin
 from betty.tests.conftest import (
     check_skip_playwright,
     check_skip_webpack_entry_point_provider,
@@ -27,7 +27,9 @@ class TestSearchUi:
         async with Project.new_isolated(extensions=[RaspberryMint]) as project:
             project.ancestry[Person].add(person)
             await generate(project)
-            async with await builtin.Builtin.new(project) as server:
+            async with await project_builtin.ProjectBuiltinServer.new(
+                project
+            ) as server:
                 yield project, server
 
     @pytest.mark.asyncio(loop_scope="session")
